+++ src/main/java/dev/tobee/telegram/request/body/ForwardMessageBody.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Optional;
 
 public record ForwardMessageBody(
-        @JsonProperty("chat_id") int chatId,
+        @JsonProperty("chat_id") long chatId,
         @JsonProperty("from_chat_id") int fromChatId,
         @JsonProperty("disable_notification")Optional<Boolean> disableNotification,
         @JsonProperty("message_id") int messageId
